Node priority type made float. Calculating time fixed.

+++ Pathfinding/Assets/Scripts/Node.cs
@@ -23,7 +23,7 @@ public class Node : IComparable<Node>
     public List<Node> neighbors = new List<Node>();
     public Node previous = null;
     public float distanceTraveled = Mathf.Infinity;
-    public int priority;
+    public float priority;
 
     public Node(int xIndex, int yIndex, NodeType nodeType)
     {
